+++ src/clients/alphaVantage/models/search/types.ts
@@ -1,0 +1,31 @@
+export interface MatchStock {
+  symbol?: string;
+  name?: string;
+  type?: string;
+  region?: string;
+  marketOpen?: string;
+  marketClose?: string;
+  timezone?: string;
+  currency?: string;
+  matchScore?: string;
+}
+
+export interface RawStock {
+  '1. symbol'?: string;
+  '2. name'?: string;
+  '3. type'?: string;
+  '4. region'?: string;
+  '5. marketOpen'?: string;
+  '6. marketClose'?: string;
+  '7. timezone'?: string;
+  '8. currency'?: string;
+  '9. matchScore'?: string;
+}
+
+export interface SearchResponse {
+  bestMatches: MatchStock[];
+}
+
+export interface RawSearchResponse {
+  bestMatches: RawStock[];
+}