+++ src/clients/alphaVantage/parsers/base.ts
@@ -1,0 +1,3 @@
+export abstract class BaseParser<M, T> {
+  public abstract Parse(data: M): T;
+}