+++ src/resolvers/search/index.ts
@@ -1,13 +1,22 @@
 // Resolvers define the technique for fetching the types defined in the
 
+import { SearchResponse } from '../../clients/alphaVantage/models';
 import { StockController } from '../../controllers/stocks/index';
-import { SearchStockResponse } from '../../controllers/stocks/models';
+
+interface ISearchArgs {
+  input: string;
+}
 
 // schema.
 export const search = {
   Query: {
-    GetStocksGQL(): SearchStockResponse {
-      return StockController.GetStocks();
+    async GetStocksGQL(
+      _: unknown,
+      args: ISearchArgs
+    ): Promise<SearchResponse | undefined> {
+      if (args.input) {
+        return await StockController.GetStocks(args.input);
+      }
     },
   },
 };