+++ CHANGELOG.md
@@ -1,5 +1,8 @@
+# Project Changelog
+
 ## Unreleased
 
+- Feature: Added the a AlphaVantage client and a single operational controller function to get stock names using a keyword. ([#1](https://github.com/josh-W42/stonks-graphwatch/compare/feature/alpha-vantage-client?expand=1))
 - Dev: Inital Commit. Created simple graphql server.
 
 ## Project Creation: April 19th 2022